// src/models/Inspection.model.ts
export enum StatusInspection {
    In_Progress = 'In_Progress',
    Completed = 'Completed',
    Canceled = 'Canceled',
    Planned = "Planned"
  }
  
  export enum TypeInspection {
    Structural = 'Structural',
    Security = 'Security',
    Installation = 'Installation',
    Finishing = 'Finishing'
  }
// ...
  export interface Projet {
    idProjet: number;
    nomProjet: string; 
  }
  
  export interface Inspection {
  
    idProjet: any;
    idINS: number;
    dateInspection: string; 
    typeInspection: TypeInspection;
    statusInspection: StatusInspection;
    nonConformities:[]
    rapportQualite?: RapportQualite; 
    projet:Projet | null | undefined;
    user: User | null | undefined;
  }
// ...
  export interface User {
    idUSER: number;
    nom: string;
    adresse: string;
    telephone: string;
    email: string;
    userRole: UserRole;
  }
  
  export enum UserRole {
    ARCHITECTE = 'ARCHITECTE',
    INGENIEUR = 'INGENIEUR',
    CLIENT = 'CLIENT',
    INSPECTEUR = 'Inspecteur'
  }
  
  
  export interface RapportQualite {
    typeInspection: any;
    statusInspection: any;
    nonConformities: boolean;
    date: string;
    statut: any;
    nonConformites: string;
    actionsCorrectives: string;
    idR: number;
    contenu: string;
    photoVideo?: string;
    dateCreation: string;
    inspectionId: number;
  }